Vendor grammar files into repo.
As noted on:
https://github.com/JustusAdam/language-haskell/issues/169

The grammar files are not vendored into the repo by default,
which is inconvenient if you want to scrape the TextMate grammars
using ordinary GET requests to GitHub.

This is the result of running `make all` at
c71c5d06452ec63c525cd6de27a16689302af5c5.